Rm test_jit.cpp
