+++ test/cpp/jit/gtest.cpp
@@ -17,12 +17,13 @@ JIT_TEST(CodeTemplate)
 JIT_TEST(ControlFlow)
 JIT_TEST(CreateAutodiffSubgraphs)
 JIT_TEST(CustomOperators)
-JIT_TEST(SchemaParser)
 JIT_TEST(Differentiate)
 JIT_TEST(DifferentiateWithRequiresGrad)
+JIT_TEST(DynamicDAG)
 JIT_TEST(FromQualString)
 JIT_TEST(InternedStrings)
 JIT_TEST(IValue)
+JIT_TEST(SchemaParser)
 JIT_TEST(TopologicalIndex)
 
 #define JIT_TEST_CUDA(name)    \
