+++ test/cpp/jit/no-gtest.cpp
@@ -15,9 +15,9 @@ std::string runJITCPPTests() {
   testControlFlow();
   testCreateAutodiffSubgraphs(out);
   testCustomOperators();
-  testSchemaParser();
   testDifferentiate(out);
   testDifferentiateWithRequiresGrad(out);
+  testDynamicDAG();
   testFromQualString();
   testFusion();
   testGraphExecutor();
@@ -25,6 +25,7 @@ std::string runJITCPPTests() {
   testInterp();
   testIValue();
   testProto();
+  testSchemaParser();
   return out.str();
 }
 } // namespace jit
