+++ test/cpp/jit/tests.h
@@ -34,6 +34,7 @@
 #include "torch/csrc/jit/autodiff.h"
 #include "torch/csrc/jit/code_template.h"
 #include "torch/csrc/jit/custom_operator.h"
+#include "torch/csrc/jit/dynamic_dag.h"
 #include "torch/csrc/jit/fusers/interface.h"
 #include "torch/csrc/jit/interned_strings.h"
 #include "torch/csrc/jit/interpreter.h"
@@ -61,6 +62,8 @@
 #include "onnx/onnx_pb.h"
 
 #include <ATen/ATen.h>
+
+#include <c10/util/Exception.h>
 
 #include <algorithm>
 #include <cstddef>
@@ -1220,6 +1223,195 @@ void testTopologicalIndex() {
   }
 }
 
+
+std::unique_ptr<detail::DynamicDAG<std::string>> newDynamicDAG() {
+  return std::unique_ptr<detail::DynamicDAG<std::string>>(new detail::DynamicDAG<std::string>());
+}
+
+void testNewVertex() {
+  auto graph = newDynamicDAG();
+  JIT_ASSERT(graph->debugNumVertices() == 0);
+  auto a = graph->newVertex("a");
+  JIT_ASSERT(graph->debugNumVertices() == 1);
+  JIT_ASSERT(a->ord == 0);
+  JIT_ASSERT(a->data.size() == 1);
+  JIT_ASSERT(a->data[0] == "a");
+  JIT_ASSERT(a->in_edges().size() == 0);
+  JIT_ASSERT(a->out_edges().size() == 0);
+  auto b = graph->newVertex("b");
+  auto c = graph->newVertex("c");
+  JIT_ASSERT(graph->debugNumVertices() == 3);
+  JIT_ASSERT(b->ord == 1);
+  JIT_ASSERT(c->ord == 2);
+}
+
+void testAddEdgeBasic() {
+  // a -> b -> c
+  // \---------^
+  auto graph = newDynamicDAG();
+  auto a = graph->newVertex("a");
+  auto b = graph->newVertex("b");
+  auto c = graph->newVertex("c");
+  graph->addEdge(a, b);
+  graph->addEdge(b, c);
+  graph->addEdge(a, c);
+  JIT_ASSERT(a->in_edges().size() == 0);
+  JIT_ASSERT(a->out_edges().size() == 2);
+  JIT_ASSERT(a->out_edges().contains(b));
+  JIT_ASSERT(a->out_edges().contains(c));
+  JIT_ASSERT(b->in_edges().size() == 1);
+  JIT_ASSERT(b->out_edges().size() == 1);
+  JIT_ASSERT(b->in_edges().contains(a));
+  JIT_ASSERT(b->out_edges().contains(c));
+  JIT_ASSERT(c->in_edges().size() == 2);
+  JIT_ASSERT(c->out_edges().size() == 0);
+  JIT_ASSERT(c->in_edges().contains(a));
+  JIT_ASSERT(c->in_edges().contains(b));
+}
+
+void testAddEdgeCycleDetection() {
+  // a -> b -> c
+  // ^---------/
+  auto graph = newDynamicDAG();
+  auto a = graph->newVertex("a");
+  auto b = graph->newVertex("b");
+  auto c = graph->newVertex("c");
+  graph->addEdge(a, b);
+  graph->addEdge(b, c);
+  bool erred = false;
+  try {
+    graph->addEdge(c, a);
+  } catch (c10::Error& err) {
+    erred = true;
+  }
+  JIT_ASSERT(erred);
+}
+
+void testAddEdgeReordersBasic() {
+  // a, b => b -> a
+  auto graph = newDynamicDAG();
+  auto a = graph->newVertex("a");
+  auto b = graph->newVertex("b");
+  JIT_ASSERT(a->ord == 0);
+  JIT_ASSERT(b->ord == 1);
+  graph->addEdge(b, a);
+  JIT_ASSERT(a->ord == 1);
+  JIT_ASSERT(b->ord == 0);
+}
+
+void testAddEdgeReordersComplicated() {
+  // a -> b  c -> d with addEdge(d, b) ==>
+  // c -> d -> a -> b
+  auto graph = newDynamicDAG();
+  auto a = graph->newVertex("a");
+  auto b = graph->newVertex("b");
+  auto c = graph->newVertex("c");
+  auto d = graph->newVertex("d");
+  graph->addEdge(a, b);
+  graph->addEdge(c, d);
+  JIT_ASSERT(a->ord == 0);
+  JIT_ASSERT(b->ord == 1);
+  JIT_ASSERT(c->ord == 2);
+  JIT_ASSERT(d->ord == 3);
+  graph->addEdge(d, a);
+  JIT_ASSERT(c->ord == 0);
+  JIT_ASSERT(d->ord == 1);
+  JIT_ASSERT(a->ord == 2);
+  JIT_ASSERT(b->ord == 3);
+  JIT_ASSERT(c->in_edges().size() == 0);
+  JIT_ASSERT(c->out_edges().size() == 1);
+  JIT_ASSERT(c->out_edges().contains(d));
+  JIT_ASSERT(d->in_edges().size() == 1);
+  JIT_ASSERT(d->out_edges().size() == 1);
+  JIT_ASSERT(d->in_edges().contains(c));
+  JIT_ASSERT(d->out_edges().contains(a));
+  JIT_ASSERT(a->in_edges().size() == 1);
+  JIT_ASSERT(a->out_edges().size() == 1);
+  JIT_ASSERT(a->in_edges().contains(d));
+  JIT_ASSERT(a->out_edges().contains(b));
+  JIT_ASSERT(b->in_edges().size() == 1);
+  JIT_ASSERT(b->out_edges().size() == 0);
+  JIT_ASSERT(b->in_edges().contains(a));
+}
+
+void testRemoveEdgeBasic() {
+  // a -> b
+  auto graph = newDynamicDAG();
+  auto a = graph->newVertex("a");
+  auto b = graph->newVertex("b");
+  graph->addEdge(a, b);
+  JIT_ASSERT(graph->debugNumVertices() == 2);
+  graph->removeEdge(a, b);
+  JIT_ASSERT(graph->debugNumVertices() == 2);
+  JIT_ASSERT(a->out_edges().size() == 0);
+  JIT_ASSERT(b->in_edges().size() == 0);
+}
+
+void testRemoveVertexBasic() {
+  // a -> b
+  auto graph = newDynamicDAG();
+  auto a = graph->newVertex("a");
+  auto b = graph->newVertex("b");
+  auto c = graph->newVertex("c");
+  graph->addEdge(a, b);
+  graph->addEdge(b, c);
+  JIT_ASSERT(graph->debugNumVertices() == 3);
+  graph->removeVertex(b);
+  JIT_ASSERT(graph->debugNumVertices() == 2);
+  JIT_ASSERT(a->out_edges().size() == 0);
+  JIT_ASSERT(c->in_edges().size() == 0);
+}
+
+void testContractEdgeBasic() {
+  // a -> b -> c -> d
+  auto graph = newDynamicDAG();
+  auto a = graph->newVertex("a");
+  auto b = graph->newVertex("b");
+  auto c = graph->newVertex("c");
+  auto d = graph->newVertex("d");
+  graph->addEdge(a, b);
+  graph->addEdge(b, c);
+  graph->addEdge(c, d);
+  graph->contractEdge(b, c);
+  JIT_ASSERT(graph->debugNumVertices() == 3);
+  JIT_ASSERT(a->out_edges().size() == 1);
+  JIT_ASSERT(d->in_edges().size() == 1);
+  JIT_ASSERT(*a->out_edges().begin() == *d->in_edges().begin());
+  auto* contracted = *a->out_edges().begin();
+  JIT_ASSERT(contracted->data.size() == 2);
+  JIT_ASSERT(contracted->data[0] == "b");
+  JIT_ASSERT(contracted->data[1] == "c");
+  JIT_ASSERT(contracted->out_edges().size() == 1);
+  JIT_ASSERT(contracted->in_edges().size() == 1);
+  JIT_ASSERT(contracted->in_edges().contains(a));
+  JIT_ASSERT(contracted->out_edges().contains(d));
+}
+
+void testContractEdgeCycleDetection() {
+  // a -> b -> c
+  // `---------^
+  // contractEdge(a, c) will cause a cycle
+  auto graph = newDynamicDAG();
+  auto a = graph->newVertex("a");
+  auto b = graph->newVertex("b");
+  auto c = graph->newVertex("c");
+  graph->addEdge(a, b);
+  graph->addEdge(b, c);
+  graph->addEdge(a, c);
+  JIT_ASSERT(!graph->contractEdge(a, c));
+}
+
+void testDynamicDAG() {
+  testNewVertex();
+  testAddEdgeBasic();
+  testAddEdgeCycleDetection();
+  testAddEdgeReordersBasic();
+  testAddEdgeReordersComplicated();
+  testRemoveEdgeBasic();
+  testRemoveVertexBasic();
+  testContractEdgeBasic();
+  testContractEdgeCycleDetection();
+}
 } // namespace
 } // namespace jit
 } // namespace torch
